fixed the searchLogic
