+++ components/RouteList/index.js
@@ -1,19 +1,45 @@
+import { useState } from "react";
 import RouteCard from "../RouteCard/index.js";
 import SearchBar from "../SearchBar/index.js";
-import { ListItem, Heading, List } from "./RouteList.styled";
+import { ListItem, List } from "./RouteList.styled";
 
 export default function RouteList({ routesData }) {
+  const [searchQuery, setSearchQuery] = useState("");
+  const [searchResults, setSearchResults] = useState([]);
+
+  function handleSearch(event) {
+    event.preventDefault();
+    const query = event.target.value;
+    setSearchQuery(query);
+
+    const results = routesData.filter((route) =>
+      route.name.toLowerCase().includes(query.toLowerCase())
+    );
+    setSearchResults(results);
+  }
   return (
     <>
-      <SearchBar />
+      <SearchBar
+        searchQuery={searchQuery}
+        handleSearch={handleSearch}
+        searchResults={searchResults}
+      />
       <List role="list">
-        {routesData.map((route) => {
-          return (
+        {searchQuery === "" ? (
+          routesData.map((route) => (
             <ListItem key={route.id}>
               <RouteCard route={route} id={route.id} />
             </ListItem>
-          );
-        })}
+          ))
+        ) : searchResults.length > 0 ? (
+          searchResults.map((route) => (
+            <ListItem key={route.id}>
+              <RouteCard route={route} id={route.id} />
+            </ListItem>
+          ))
+        ) : (
+          <ListItem>No matching route found.</ListItem>
+        )}
       </List>
     </>
   );
