+++ components/SearchBar/index.js
@@ -1,23 +1,6 @@
 import styled from "styled-components";
-import { useState } from "react";
-import RouteCard from "../RouteCard";
-import { List, ListItem } from "../RouteList/RouteList.styled";
-import { routesData } from "@/routesData";
 
-export default function SearchBar() {
-  const [searchQuery, setSearchQuery] = useState("");
-  const [searchResults, setSearchResults] = useState([]);
-
-  function handleSearch(event) {
-    event.preventDefault();
-    const query = event.target.value;
-    setSearchQuery(query);
-
-    const results = routesData.filter((route) =>
-      route.name.toLowerCase().includes(query.toLowerCase())
-    );
-    setSearchResults(results);
-  }
+export default function SearchBar({searchQuery,handleSearch,searchResults  }) {
 
   return (
     <SearchContainer>
@@ -29,17 +12,6 @@ export default function SearchBar() {
         name="search"
         onChange={handleSearch}
       />
-      {searchResults.length > 0 && (
-        <List>
-          {searchResults.map((result) => {
-            return (
-              <ListItem key={result.id}>
-                <RouteCard route={result} id={result.id} />
-              </ListItem>
-            );
-          })}
-        </List>
-      )}
     </SearchContainer>
   );
 }
