+++ pages/_app.js
@@ -1,14 +1,12 @@
 import NavigationBar from "@/components/NavigationBar";
 import GlobalStyle from "../styles";
 
-
 export default function App({ Component, pageProps }) {
   return (
     <>
       <GlobalStyle />
       <Component {...pageProps} />
       <NavigationBar />
-      
     </>
   );
 }
